Add a startDaemon flag
This shouldn't change the default behaviour of the library (i.e.
when no startDaemon flag is passed, the library still downloads
and runs the IPFS daemon itself). However, when the startDaemon
flag is set to `false`, a daemon is not started and the specified
node is assumed to be running.